backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.4.0
psycopg==3.1.18
Pillow>=10.2.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0 
djangorestframework-simplejwt==5.4.0
razorpay==1.3.0
setuptools==68.2.2
cloudinary
django-cloudinary-storage